pyproject.toml
[project]
name = "music-manager"
version = "1.0.0"
description = ""
authors = [
    {name = "rboiko",email = "rboiko@deloittece.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aiohttp (>=3.11.11,<4.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
